Cleanup

+++ tt_metal/hw/inc/api/dataflow/dataflow_api.h
@@ -27,7 +27,6 @@
 #include "api/compile_time_args.h"
 #include "hostdev/dev_msgs.h"
 #include "api/tensor/tensor_accessor.h"
-#include "api/debug/dprint.h"
 #include "tools/profiler/kernel_profiler.hpp"
 #include "internal/debug/sanitize.h"
 
@@ -1468,8 +1467,8 @@ FORCE_INLINE void noc_async_write_shard(
  */
 // clang-format on
 template <ProgrammableCoreType type = ProgrammableCoreType::TENSIX>
-FORCE_INLINE uintptr_t get_semaphore(uint32_t semaphore_id) {
-    return (uintptr_t)sem_l1_base[static_cast<int>(type)] + semaphore_id * L1_ALIGNMENT;
+FORCE_INLINE uint32_t get_semaphore(uint32_t semaphore_id) {
+    return (uint32_t)sem_l1_base[static_cast<int>(type)] + semaphore_id * L1_ALIGNMENT;
 }
 
 // clang-format off
